// src/core/runtime/SdkDefinitions.generated.ts
export const kSdkTypes: readonly SdkTypeDefinition[] = [
  { name: "AirDirection", baseType: "int", arrayExtent: 0 },
  { name: "CenterLineGroup", baseType: "int", arrayExtent: 0 },
  { name: "CpInstType", baseType: "int", arrayExtent: 0 },
  { name: "DimMode", baseType: "int", arrayExtent: 0 },
  { name: "ElementRole", baseType: "int", arrayExtent: 0 },
  { name: "ElementType", baseType: "int", arrayExtent: 0 },
  { name: "ElemsType", baseType: "int", arrayExtent: 0 },
  { name: "ElevRef", baseType: "int", arrayExtent: 0 },
  { name: "ErrorStatus", baseType: "int", arrayExtent: 0 },
  { name: "FLM3Geo::ErrorStatus", baseType: "int", arrayExtent: 0 },
  { name: "FLM3Geo::geoRepMode", baseType: "int", arrayExtent: 0 },
  { name: "FLM3Geo::primitiveMode", baseType: "int", arrayExtent: 0 },
  { name: "FdGeoParamType", baseType: "int", arrayExtent: 0 },
  { name: "FdPropStrType", baseType: "int", arrayExtent: 0 },
  { name: "FdView2d", baseType: "int", arrayExtent: 0 },
  { name: "FlexProfile", baseType: "int", arrayExtent: 0 },
  { name: "GeometryType", baseType: "int", arrayExtent: 0 },
  { name: "GroupType", baseType: "int", arrayExtent: 0 },
  { name: "InsulationType", baseType: "int", arrayExtent: 0 },
  { name: "LibType", baseType: "int", arrayExtent: 0 },
  { name: "LinkType", baseType: "int", arrayExtent: 0 },
  { name: "PLossType", baseType: "int", arrayExtent: 0 },
  { name: "PipeType", baseType: "int", arrayExtent: 0 },
  { name: "PowerTableType", baseType: "int", arrayExtent: 0 },
  { name: "SystemType", baseType: "int", arrayExtent: 0 },
  { name: "TechType", baseType: "int", arrayExtent: 0 },
  { name: "ads_point", baseType: "double", arrayExtent: 3 },
  { name: "ads_real", baseType: "double", arrayExtent: 0 },
  { name: "cpFlowType", baseType: "int", arrayExtent: 0 },
  { name: "cpHeatFunc", baseType: "int", arrayExtent: 0 },
  { name: "fdCircSymbType", baseType: "int", arrayExtent: 0 },
  { name: "geoRepMode", baseType: "int", arrayExtent: 0 },
  { name: "hlr_curve_type", baseType: "int", arrayExtent: 0 },
  { name: "line_type", baseType: "int", arrayExtent: 0 },
  { name: "primitiveMode", baseType: "int", arrayExtent: 0 },
];